added better comments
