+++ app/user/model.ts
@@ -1,6 +1,12 @@
-import { getDataFromFile, writeData } from "../utils";
-import { Role } from "../types";
+import { getDataFromFile, writeData } from "../core/utils/utils";
+import { Role } from "./types";
 
+/**
+ * Creates an class UserModel.
+ *
+ * @constructor
+ * @param {object} init This holds the data about a user. this holds the data like id, firstName, lastName and more.
+ */
 export class UserModel {
   id?: string;
   firstName: string = "";
@@ -15,6 +21,11 @@ export class UserModel {
     Object.assign(this, init);
   }
 
+  /**
+   * Saves a new user to database
+   *
+   * @this {UserModel}
+   */
   save() {
     getDataFromFile("user.json", (users: UserModel[]) => {
       this.id = Date.now().toString();
